+++ packages/pn-commons/src/components/CopyToClipboard.tsx
@@ -1,31 +1,39 @@
-import { useEffect, useState } from 'react';
+import { ComponentPropsWithoutRef, useEffect, useState } from 'react';
 
 import CheckIcon from '@mui/icons-material/Check';
 import ContentCopyIcon from '@mui/icons-material/ContentCopy';
-import { Button, Link, SxProps, Theme, Tooltip } from '@mui/material';
+import { Button, SxProps, Theme, Tooltip } from '@mui/material';
+import { ButtonNaked } from '@pagopa/mui-italia';
 
 import { useIsMobile } from '../hooks';
+
+type AllowedSlots = typeof Button | typeof ButtonNaked;
+type AllowedSlotProps = ComponentPropsWithoutRef<AllowedSlots>;
 
 type Props = {
   /** callback used to retrieve the text to be copied */
   getValue: () => string;
   /** an optional text to be displayed near the "copy to clipboard" icon */
   text?: string;
-  textPosition?: "start" | "end";
+  textPosition?: 'start' | 'end';
   tooltipMode?: boolean;
   tooltip?: string;
   tooltipBefore?: string;
   disabled?: boolean;
+  slot?: AllowedSlots;
+  slotProps?: AllowedSlotProps;
 };
 
 const CopyToClipboard: React.FC<Props> = ({
   getValue,
   text,
-  textPosition = "end",
+  textPosition = 'end',
   tooltipMode,
   tooltip = '',
   tooltipBefore = '',
   disabled = false,
+  slot,
+  slotProps = {},
 }) => {
   const padding = tooltipMode ? 0 : undefined;
   const alertButtonStyle: SxProps<Theme> = useIsMobile()
@@ -55,14 +63,16 @@ const CopyToClipboard: React.FC<Props> = ({
     }
   };
 
+  const SlotComponent = slot || Button;
+
   return (
-    <Button
-      component={Link}
+    <SlotComponent
       color="primary"
       sx={{ ...alertButtonStyle }}
       onClick={doCopyToClipboard}
       disabled={disabled}
       aria-label={copied ? tooltip : tooltipBefore}
+      {...slotProps}
     >
       {textPosition === 'start' && text}
       {copied && (
@@ -72,7 +82,7 @@ const CopyToClipboard: React.FC<Props> = ({
       )}
       {!copied && <ContentCopyIcon fontSize="small" sx={{ m: '5px' }} />}
       {textPosition === 'end' && text}
-    </Button>
+    </SlotComponent>
   );
 };