+++ packages/pn-personafisica-webapp/src/components/Contacts/PecContactWizard.tsx
@@ -72,9 +72,12 @@ const PecContactWizard: React.FC<Props> = ({
     },
   });
 
-  const handleChangeTouched = async (e: ChangeEvent) => {
+  const handleChangeTouched = async (e: ChangeEvent<HTMLInputElement>) => {
     if (e.target.id === 'disclaimer') {
-      PFEventStrategyFactory.triggerEvent(PFEventsType.SEND_ADD_SERCQ_SEND_PEC_TOS_ACCEPTED);
+      const event = e.target.checked
+        ? PFEventsType.SEND_ADD_SERCQ_SEND_PEC_TOS_ACCEPTED
+        : PFEventsType.SEND_ADD_SERCQ_SEND_PEC_TOS_DISMISSED;
+      PFEventStrategyFactory.triggerEvent(event);
     }
 
     formik.handleChange(e);
@@ -129,8 +132,11 @@ const PecContactWizard: React.FC<Props> = ({
 
   const handleSubmitForm = async () => {
     const errors = formik.errors;
+    const pecValidation = !formik.values.pec ? 'missing' : errors?.pec ? 'invalid' : 'valid';
     if (errors?.pec) {
-      PFEventStrategyFactory.triggerEvent(PFEventsType.SEND_ADD_SERCQ_SEND_PEC_MISSING);
+      PFEventStrategyFactory.triggerEvent(PFEventsType.SEND_ADD_SERCQ_SEND_PEC_ERROR, {
+        pec_validation: pecValidation,
+      });
     }
 
     if (errors?.disclaimer) {
@@ -138,7 +144,7 @@ const PecContactWizard: React.FC<Props> = ({
     }
 
     PFEventStrategyFactory.triggerEvent(PFEventsType.SEND_ADD_SERCQ_SEND_PEC_START_ACTIVATION, {
-      pec_validation: !formik.values.pec ? 'missing' : errors?.pec ? 'invalid' : 'valid',
+      pec_validation: pecValidation,
       tos_validation: errors?.disclaimer ? 'missing' : 'valid',
     });