+++ packages/pn-personafisica-webapp/src/utility/AppError/__test__/PFAppErrorFactory.test.ts
@@ -1,14 +1,19 @@
-import { AppError, ServerResponseError } from '@pagopa-pn/pn-commons';
+import { AppError, ServerResponseError, UnknownAppError } from '@pagopa-pn/pn-commons';
 
-import { PFAppErrorFactory } from '../PFAppErrorFactory';
-import { ServerResponseErrorCode } from '../types';
+import { DeliveryFileInfoNotFoundAppError } from '../DeliveryFileInfoNotFoundAppError';
+import { DeliveryMandateNotFoundAppError } from '../DeliveryMandateNotFoundAppError';
+import { DeliveryNotificationWithoutPaymentAttachmentAppError } from '../DeliveryNotificationWithoutPaymentAttachmentAppError';
+import { DeliveryPushFileNotFoundAppError } from '../DeliveryPushFileNotFoundAppError';
 import { GenericInvalidParameterPatternAppError } from '../GenericInvalidParameterPatternAppError';
+import { InvalidBodyAppError } from '../InvalidBodyAppError';
 import { MandateAlreadyExistsAppError } from '../MandateAlreadyExistsAppError';
 import { MandateDelegateHimselfAppError } from '../MandateDelegateHimselfAppError';
 import { MandateInvalidVerificationCodeAppError } from '../MandateInvalidVerificationCodeAppError';
 import { MandateNotAcceptableAppError } from '../MandateNotAcceptableAppError';
 import { MandateNotFoundAppError } from '../MandateNotFoundAppError';
+import { PFAppErrorFactory } from '../PFAppErrorFactory';
 import { UserAttributesInvalidVerificationCodeAppError } from '../UserAttributesInvalidVerificationCodeAppError';
+import { ServerResponseErrorCode } from '../types';
 
 class PFAppErrorFactoryForTest extends PFAppErrorFactory {
   constructor(translateFunction: (path: string, ns: string) => string) {
@@ -71,5 +76,45 @@ describe('Test PFAppErrorFactory', () => {
       code: ServerResponseErrorCode.PN_GENERIC_INVALIDPARAMETER_PATTERN,
     });
     expect(errorClass).toBeInstanceOf(GenericInvalidParameterPatternAppError);
+  });
+
+  it('Should return DeliveryMandateNotFoundAppError', () => {
+    const errorClass = errorFactory.getCustomErrorForTest({
+      code: ServerResponseErrorCode.PN_DELIVERY_MANDATENOTFOUND,
+    });
+    expect(errorClass).toBeInstanceOf(DeliveryMandateNotFoundAppError);
+  });
+
+  it('Should return InvalidBodyAppError', () => {
+    const errorClass = errorFactory.getCustomErrorForTest({
+      code: ServerResponseErrorCode.PN_INVALID_BODY,
+    });
+    expect(errorClass).toBeInstanceOf(InvalidBodyAppError);
+  });
+
+  it('Should return DeliveryFileInfoNotFoundAppError', () => {
+    const errorClass = errorFactory.getCustomErrorForTest({
+      code: ServerResponseErrorCode.PN_DELIVERY_FILEINFONOTFOUND,
+    });
+    expect(errorClass).toBeInstanceOf(DeliveryFileInfoNotFoundAppError);
+  });
+
+  it('Should return DeliveryNotificationWithoutPaymentAttachmentAppError', () => {
+    const errorClass = errorFactory.getCustomErrorForTest({
+      code: ServerResponseErrorCode.PN_DELIVERY_NOTIFICATIONWITHOUTPAYMENTATTACHMENT,
+    });
+    expect(errorClass).toBeInstanceOf(DeliveryNotificationWithoutPaymentAttachmentAppError);
+  });
+
+  it('Should return DeliveryPushFileNotFoundAppError', () => {
+    const errorClass = errorFactory.getCustomErrorForTest({
+      code: ServerResponseErrorCode.PN_DELIVERYPUSH_FILE_NOT_FOUND,
+    });
+    expect(errorClass).toBeInstanceOf(DeliveryPushFileNotFoundAppError);
+  });
+
+  it('Should return UnknownAppError', () => {
+    const errorClass = errorFactory.getCustomErrorForTest({ code: 'UNKNOWN_CODE' });
+    expect(errorClass).toBeInstanceOf(UnknownAppError);
   });
 });