+++ packages/pn-personafisica-webapp/src/utility/MixpanelUtils/Strategies/KoErrorStrategy.ts
@@ -7,11 +7,20 @@ import {
 } from '@pagopa-pn/pn-commons';
 
 export class KoErrorStrategy implements EventStrategy {
-  performComputations(): TrackedEvent {
+  performComputations(details?: { [key: string]: string }): TrackedEvent {
+    if (!details) {
+      return {
+        [EventPropertyType.TRACK]: {
+          event_category: EventCategory.KO,
+          event_type: EventAction.ERROR,
+        },
+      };
+    }
     return {
       [EventPropertyType.TRACK]: {
         event_category: EventCategory.KO,
         event_type: EventAction.ERROR,
+        ...details,
       },
     };
   }