+++ packages/pn-personagiuridica-webapp/src/components/Contacts/EmailContactItem.tsx
@@ -16,6 +16,7 @@ import DigitalContact from './DigitalContact';
 import ExistingContactDialog from './ExistingContactDialog';
 import InformativeDialog from './InformativeDialog';
 import SmsContactItem from './SmsContactItem';
+import SpecialContacts from './SpecialContacts';
 
 enum ModalType {
   EXISTING = 'existing',
@@ -31,6 +32,8 @@ const EmailContactItem: React.FC = () => {
     defaultPECAddress,
     defaultEMAILAddress,
     defaultSMSAddress,
+    specialAddresses,
+    specialEMAILAddresses,
     addresses,
   } = useAppSelector(contactsSelectors.selectAddresses);
   const digitalContactRef = useRef<{ toggleEdit: () => void; resetForm: () => Promise<void> }>({
@@ -44,6 +47,10 @@ const EmailContactItem: React.FC = () => {
   const hasCourtesyAddresses =
     addresses.filter((addr) => addr.addressType === AddressType.COURTESY).length > 0;
 
+  const showSpecialContactsSection = !defaultSMSAddress
+    ? specialAddresses.filter((addr) => addr.addressType === AddressType.COURTESY).length > 0
+    : specialEMAILAddresses.length > 0;
+
   const [modalOpen, setModalOpen] = useState<ModalType | null>(null);
   // currentAddress is needed to store what address we are creating/editing/removing
   // because this variable isn't been used to render, we can use useRef
@@ -53,6 +60,7 @@ const EmailContactItem: React.FC = () => {
   const dispatch = useAppDispatch();
 
   const currentValue = defaultEMAILAddress?.value ?? '';
+  const blockDelete = specialEMAILAddresses.length > 0;
 
   const handleSubmit = (value: string) => {
     // eslint-disable-next-line functional/immutable-data
@@ -146,6 +154,9 @@ const EmailContactItem: React.FC = () => {
   };
 
   const getRemoveModalTitle = () => {
+    if (blockDelete) {
+      return t(`courtesy-contacts.block-remove-email-title`, { ns: 'recapiti' });
+    }
     if (defaultSERCQ_SENDAddress) {
       return t(`courtesy-contacts.remove-email-title-dod-enabled`, {
         ns: 'recapiti',
@@ -155,6 +166,9 @@ const EmailContactItem: React.FC = () => {
   };
 
   const getRemoveModalMessage = () => {
+    if (blockDelete) {
+      return t(`courtesy-contacts.block-remove-email-message`, { ns: 'recapiti' });
+    }
     if (defaultSERCQ_SENDAddress) {
       return (
         <Trans
@@ -251,12 +265,17 @@ const EmailContactItem: React.FC = () => {
           {t('courtesy-contacts.email-filled-description', { ns: 'recapiti' })}
         </Typography>
       )}
-
       {!defaultSMSAddress && (
         <>
           <Divider sx={{ mt: 3, mb: 3 }} />
           <SmsContactItem />
         </>
+      )}
+      {showSpecialContactsSection && (
+        <SpecialContacts
+          addressType={AddressType.COURTESY}
+          channelType={!defaultSMSAddress ? undefined : ChannelType.EMAIL}
+        />
       )}
       <ExistingContactDialog
         open={modalOpen === ModalType.EXISTING}
@@ -290,6 +309,7 @@ const EmailContactItem: React.FC = () => {
               : undefined,
           },
         }}
+        blockDelete={blockDelete}
       />
       <InformativeDialog
         open={modalOpen === ModalType.INFORMATIVE}